:recycle: 유저 전체 조회, 유저 닉네임 조회 위치 변경
