+++ src/main/java/com/maple/volunteer/service/user/UserService.java
@@ -250,67 +250,6 @@ public class UserService {
 //        return commonService.successResponse(SuccessCode.NEW_USER_SUCCESS.getDescription(), HttpStatus.OK, newUserDto);
 //    }
 
-    // 유저 조회 (true, false 모두)
-    public CommonResponseDto<Object> allUserInquiry (int page, int size, String sortBy ) {
-
-        // 페이지, 요소 개수, 정렬
-        PageRequest pageable = PageRequest.of(page -1, size, Sort.by(sortBy).descending());
-
-        // 페이지 값 받아오기
-        Page<User> data = userRepository.findAll(pageable);
-        if (data.isEmpty()) throw new NotFoundException(ErrorCode.USER_NOT_FOUND);
-
-        // userDtoList 생성
-        List<User> userList = data.getContent();
-        List<UserDto> userDtoList = new ArrayList<>();
-
-        for (User user: userList) {
-            UserDto userDto = UserDto.builder()
-                    .email(user.getEmail())
-                    .name(user.getName())
-                    .nickname(user.getNickname())
-                    .phoneNumber(user.getPhoneNumber())
-                    .profileImg(user.getProfileImg())
-                    .build();
-            userDtoList.add(userDto);
-        }
-
-        // pagination 설정
-        PaginationDto paginationDto = PaginationDto.builder()
-                .totalPages(data.getTotalPages())
-                .totalElements(data.getTotalElements())
-                .pageNo(data.getNumber())
-                .isLastPage(data.isLast())
-                .build();
-
-        // AllUserListDto 반환
-        AllUserListDto allUserListDto = AllUserListDto.builder()
-                .userList(userDtoList)
-                .paginationDto(paginationDto)
-                .build();
-
-        return commonService.successResponse(SuccessCode.ALL_USER_INQUIRY_SUCCESS.getDescription(), HttpStatus.OK, allUserListDto);
-    }
-
-    // 닉네임으로 회원 조회 (true, false 모두)
-    public CommonResponseDto<Object> userInquiryByNickname (String nickname) {
-
-        User user = userRepository.findByNickname(nickname)
-                .orElseThrow(() -> new NotFoundException(ErrorCode.USER_NOT_FOUND));
-
-        UserDto userDto = UserDto.builder()
-                .name(user.getName())
-                .provider(user.getProvider())
-                .phoneNumber(user.getPhoneNumber())
-                .nickname(user.getNickname())
-                .email(user.getEmail())
-                .profileImg(user.getProfileImg())
-                .isDeleted(user.isDelete())
-                .build();
-
-        return commonService.successResponse(SuccessCode.VIEW_USERINFO_SUCCESS.getDescription(), HttpStatus.OK, userDto);
-    }
-
     public CommonResponseDto<Object> nicknameCheck(CheckDto checkDto) {
         Optional<User> userOptional = userRepository.findNickname(checkDto.getCheck());
         if(userOptional.isEmpty()){
